requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyYAML==6.0.1
Flask-WTF==1.2.1
Flask-Compress==1.14